rename line_localization to hybrid_localization
