+++ limap/_limap/bindings.cc
@@ -45,7 +45,7 @@ void bind_features(py::module &);
 namespace limap {
 
 PYBIND11_MODULE(_limap, m) {
-  m.doc() = "Multi-view Line Triangulation and Refinement";
+  m.doc() = "A toolbox for mapping and localization with line features";
 #ifdef VERSION_INFO
   m.attr("__version__") = py::str(VERSION_INFO);
 #else
