+++ limap/estimators/absolute_pose/joint_pose_estimator.h
@@ -4,14 +4,14 @@
 #include "_limap/helpers.h"
 #include "base/camera.h"
 #include "base/linetrack.h"
-#include "optimize/line_localization/lineloc.h"
+#include "optimize/hybrid_localization/hybrid_localization.h"
 #include "util/types.h"
 
 #include <RansacLib/ransac.h>
 
 namespace limap {
 
-using namespace optimize::line_localization;
+using namespace optimize::hybrid_localization;
 
 namespace estimators {
