+++ limap/optimize/bindings.cc
@@ -9,7 +9,7 @@
 
 #include "optimize/global_pl_association/bindings.cc"
 #include "optimize/hybrid_bundle_adjustment/bindings.cc"
-#include "optimize/line_localization/bindings.cc"
+#include "optimize/hybrid_localization/bindings.cc"
 #include "optimize/line_refinement/bindings.cc"
 
 namespace py = pybind11;
@@ -19,13 +19,13 @@ namespace limap {
 void bind_line_refinement(py::module &m);
 void bind_hybrid_bundle_adjustment(py::module &m);
 void bind_global_pl_association(py::module &m);
-void bind_line_localization(py::module &m);
+void bind_hybrid_localization(py::module &m);
 
 void bind_optimize(py::module &m) {
   bind_line_refinement(m);
   bind_hybrid_bundle_adjustment(m);
   bind_global_pl_association(m);
-  bind_line_localization(m);
+  bind_hybrid_localization(m);
 }
 
 } // namespace limap
