+++ limap/optimize/hybrid_localization/CMakeLists.txt
@@ -1,0 +1,9 @@
+set(FOLDER_NAME "hybrid_localization")
+
+LIMAP_ADD_SOURCES(
+    bindings.cc
+    cost_functions.h
+    hybrid_localization.h hybrid_localization.cc
+    hybrid_localization_config.h
+)
+
