+++ limap/optimize/hybrid_localization/hybrid_localization.cc
@@ -1,6 +1,6 @@
-#include "optimize/line_localization/lineloc.h"
+#include "optimize/hybrid_localization/hybrid_localization.h"
 #include "ceresbase/parameterization.h"
-#include "optimize/line_localization/cost_functions.h"
+#include "optimize/hybrid_localization/cost_functions.h"
 
 #include <colmap/optim/bundle_adjustment.h>
 #include <colmap/util/logging.h>
@@ -11,7 +11,7 @@ namespace limap {
 
 namespace optimize {
 
-namespace line_localization {
+namespace hybrid_localization {
 
 void LineLocEngine::ParameterizeCamera() {
   double *kvec_data = cam.Params().data();
@@ -134,7 +134,7 @@ void JointLocEngine::AddResiduals() {
   }
 }
 
-} // namespace line_localization
+} // namespace hybrid_localization
 
 } // namespace optimize
