--- limap/optimize/line_localization/CMakeLists.txt
@@ -1,9 +1,0 @@
-set(FOLDER_NAME "line_localization")
-
-LIMAP_ADD_SOURCES(
-    bindings.cc
-    cost_functions.h
-    lineloc.h lineloc.cc
-    lineloc_config.h
-)
-
