+++ runners/cambridge/localization.py
@@ -20,7 +20,7 @@ from utils import (
     undistort_and_resize,
 )
 
-import limap.runners as _runners
+import limap.runners as runners
 import limap.util.config as cfgutils
 import limap.util.io as limapio
 
@@ -116,7 +116,7 @@ def parse_config():
 
 def main():
     cfg, args = parse_config()
-    cfg = _runners.setup(cfg)
+    cfg = runners.setup(cfg)
     scene_id = os.path.basename(cfg["vsfm_path"])
 
     # outputs is for localization-related results
@@ -185,7 +185,7 @@ def main():
     finaltracks_dir = os.path.join(cfg["output_dir"], "finaltracks")
     if not cfg["skip_exists"] or not os.path.exists(finaltracks_dir):
         logger.info("Running LIMAP triangulation...")
-        linetracks_db = _runners.line_triangulation(
+        linetracks_db = runners.line_triangulation(
             cfg, imagecols_train, neighbors=neighbors, ranges=ranges
         )
     else:
@@ -222,7 +222,7 @@ def main():
         hloc_logs = pickle.load(f)
     point_correspondences = {}
     for qid in query_ids:
-        p2ds, p3ds, inliers = _runners.get_hloc_keypoints_from_log(
+        p2ds, p3ds, inliers = runners.get_hloc_keypoints_from_log(
             hloc_logs, hloc_name_dict[qid], ref_sfm
         )
         point_correspondences[qid] = {
@@ -232,7 +232,7 @@ def main():
         }
 
     # can return final_poses
-    _runners.line_localization(
+    runners.hybrid_localization(
         cfg,
         imagecols_train,
         imagecols_query,
